Adding nullifiedScore to metrics comparison (#18)
* feat: Adding nullifiedScore to metrics comparison

nullifiedScore is one of the reasons to treat one evidence string as invalid. Although this is a rather small proportion of the invalid data, we are now extracting metrics for them

* refactor: Extracted methods to make main smaller and added documentation

* fix: Expose the invalid due to null score comparison

* build: Update Pandas Styler syntax

* chore: Set GCSFS version to 2022.7.1

* build: Protobuf version pinned to 3.20 in the requirements

The build was failing because an incompatibility between Python and Protobuf. This is an installed dependency of some of the packages. By pinning the version the build no longer fails.

* style: Tidy project structure up

* style: Change src folders naming convention to use `_`

If the folder has a `-`, importing the functions from a module would fail.

* docs: Adapt docs and source to the new project structure

* feat(configuration): First iteration of the metric_calculation config

* fix: Typo

* feat: Change `read_path_if_provided` to handle GCS files

Local files are no longer expected

* feat: Parse config file and adapt source to remote files

This drops the argument parsing approach. Now the metrics generation script is parametrised from a config. On top of that, the script has been accommodated to expect remote files and not work with local files

2614, 2613

* build: Add hydra-core to requirements

* refactor: Delete Procfile

This file was necessary when the app was hosted in Heroku. Streamlit Cloud is used now

* refactor: Easier spark initialisation, added get_cwd()

* fix: Internal packages are imported from the root directory

* fix(configuration): Correct failed evidence path

* refactor: Remove complexity in read_path_if_provided

* feat: Add pathlib to requirements

* fix: Correct DataFrame import

DataFrame in this case is not only used to do the type checking

* docs: Metric calculation instructions updated

* fix: Metrics runId set to 28.08.1

* refactor: Delete gold standard files

These files are now available at gs://otar000-evidence_input/release-metrics/gold-standard

* docs: set up dataproc instructions as a single node machine

as proposed by Daniel

* build(configuration): config to extract 22.09.1 metrics

* chore(configuration): adapt config to use 22.09 first pipeline run

* chore(configuration): configuration for 22.09.3 metrics